requirements.txt
pandas
numpy
matplotlib
yfinance
requests
beautifulsoup4
nltk
tqdm
lxml
pyyaml
anthropic>=0.9.0
tiktoken
tensorflow
scikit-learn